Abstracted the FAB
